Create 2d-Chunk-Renderer Unity2D
When a player with a connected camera doesn't see any sprites that are outside the camera boundaries, they are deactivated. But as soon as he moves in any direction, these deactivated sprites, will activate 0.4 hundredths of a second in advance before the player really sees them, the same works the other way when the player leaves some sprites they will deactivate only after 0.4 seconds as well. In addition, I added a periodic check by the script for newly added sprites, so that it won't happen that when I add many new objects with sprites they will be missed. The script has a list through which I can manually add objects, but it exists so in addition, the script itself collects every object that has a “Sprite Renderer”.